fix setBitsTo prototype

+++ src/emulator/utils.h
@@ -33,7 +33,7 @@ bool getBitAt(uint64_t n, int k);
 uint64_t getBitsAt(uint64_t n, int k, int l);
 
 // Sets clearsize bits starting from kth position of n to new to size copysize.
-uint64_t setBitsTo(uint64_t n, int k, int new, int l);
+uint64_t setBitsTo(uint64_t n, int k, uint64_t new, int l);
 
 // Sign extends n at lth bit
 uint64_t extendBits(uint64_t n, int l);
